Use traits to implement the abstract Command class
