+++ src/Command/Command.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Concrete\Console\Command;
 
-use Concrete\Console\Application;
 use Concrete\Console\Concrete\Connection\ApplicationEnabledConnectionInterface;
 use Concrete\Console\Concrete\Connection\ConnectionAwareInterface;
 use Concrete\Console\Concrete\Connection\ConnectionAwareTrait;
@@ -23,34 +22,14 @@ abstract class Command implements
     InstallationAwareInterface
 {
     use ContainerAwareTrait;
+    use OutputStyleAwareTrait;
+    use ConsoleAwareTrait;
     use ConnectionAwareTrait;
     use InstallationAwareTrait;
 
-    /** @var OutputStyle */
-    protected $output;
-
-    /** @var Application */
-    protected $console;
-
     /**
-     * @param OutputStyle $outputStyle
-     * @return void
+     * @throws \Concrete\Console\Exception\Installation\VersionMismatch
      */
-    public function setOutputStyle(OutputStyle $outputStyle): void
-    {
-        $this->output = $outputStyle;
-    }
-
-    public function setConsole(Application $application): void
-    {
-        $this->console = $application;
-    }
-
-    protected function getConsole(): Application
-    {
-        return $this->console;
-    }
-
     public function getApplication(): \Concrete\Core\Application\Application
     {
         $connection = $this->getConnection();
